python/pyproject.toml
[tool.poetry]
name = "merkle-claim-arcadia"
version = "0.1.0"
description = "ad-hoc scripts to simplify claiming merkle rewards for arcadia accounts"
authors = ["j-vp <jasper@arcadia.finance>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
web3 = "^6.20.1"
python-dotenv = "^1.0.1"
multicall = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
